+++ nextgisweb/pyramid/__init__.py
@@ -1,4 +1,6 @@
+import logging
 import os.path
+from os import environ
 from datetime import datetime as dt, timedelta
 from pkg_resources import resource_filename
 
@@ -21,8 +23,11 @@ from .util import (
 from .model import Base, Session, SessionStore
 from .session import WebSession
 from .command import ServerCommand, AMDPackagesCommand  # NOQA
+from .util import _
 
 __all__ = ['viewargs', 'WebSession']
+
+logger = logging.getLogger(__name__)
 
 
 class PyramidComponent(Component):
@@ -58,6 +63,38 @@ class PyramidComponent(Component):
         view.setup_pyramid(self, config)
         api.setup_pyramid(self, config)
         uac.setup_pyramid(self, config)
+
+        try:
+            import uwsgi
+            lunkwill_rpc = b'lunkwill' in uwsgi.rpc_list()
+        except ImportError:
+            uwsgi = None
+            lunkwill_rpc = False
+
+        if self.options['lunkwill.enabled'] is None:
+            self.options['lunkwill.enabled'] = lunkwill_rpc
+       
+        if self.options['lunkwill.enabled']:
+            if self.options['lunkwill.host'] is None:
+                self.options['lunkwill.host'] = environ.get(
+                    'LUNKWILL_HOST', '127.0.0.1')
+
+            if self.options['lunkwill.port'] is None:
+                self.options['lunkwill.port'] = int(environ.get(
+                    'LUNKWILL_PORT', '8042'))
+
+            logger.debug(
+                "Lunkwill extension enabled: %s:%d",
+                self.options['lunkwill.host'],
+                self.options['lunkwill.port'])
+            if uwsgi is None:
+                raise RuntimeError("Lunkwill requires uWSGI stack loaded")
+            if not lunkwill_rpc:
+                raise RuntimeError("Lunkwill RPC missing in uWSGI stack")
+            from . import lunkwill
+            lunkwill.setup_pyramid(self, config)
+        else:
+            logger.debug("Lunkwill extension disabled")
 
     def client_settings(self, request):
         result = dict()
@@ -95,6 +132,16 @@ class PyramidComponent(Component):
 
         logger.info("Deleted: %d sessions", deleted_sessions)
 
+    def sys_info(self):
+        try:
+            import uwsgi
+            yield ("uWSGI", uwsgi.version.decode())
+        except ImportError:
+            pass
+
+        lunkwill = self.options['lunkwill.enabled']
+        yield ("Lunkwill", _("Enabled") if lunkwill else _("Disabled"))
+
     def backup_configure(self, config):
         super().backup_configure(config)
         config.exclude_table_data('public', Session.__tablename__)
@@ -124,4 +171,8 @@ class PyramidComponent(Component):
         Option('debugtoolbar.hosts'),
 
         Option('legacy_locale_switcher', bool, default=False),
+
+        Option('lunkwill.enabled', bool, default=None),
+        Option('lunkwill.host', str, default=None),
+        Option('lunkwill.port', int, default=None),
     )) + uacompat.option_annotations