+++ nextgisweb/resource/amd/ngw-resource/CompositeWidget.js
@@ -181,7 +181,7 @@ define([
             return deferred;
         },
 
-        serialize: function () {
+        serialize: function (lunkwill) {
             var promises = [];
             var data = { resource: {} };
 
@@ -191,7 +191,7 @@ define([
             }
 
             array.forEach(this.members, function (member) {
-                promises.push(when(member.serialize(data)));
+                promises.push(when(member.serialize(data, lunkwill)));
             });
 
             return all(promises).then(function () {
@@ -216,12 +216,14 @@ define([
                 function /* callback */ (success) {
                     if (success) {
                         console.debug("Validation completed with success");
-                        widget.serialize().then(
+                        var lunkwill = new api.LunkwillParam();
+                        widget.serialize(lunkwill).then(
                             function /* callback */ (data) {
                                 console.debug("Serialization completed");
                                 api.request(args.url, {
                                     method: args.method,
                                     json: data,
+                                    lunkwill: lunkwill,
                                 }).then(
                                     function /* callback */ (response) {
                                         console.debug("REST API request completed");
@@ -289,6 +291,7 @@ define([
             this.storeRequest({
                 url: api.routeURL('resource.collection'),
                 method: "POST",
+                lunkwill: "true",
             }).then(
                 /* callback */ lang.hitch(this, function (data) {
                     if (edit) {